07Metadata/requirements.txt
mysql-connector-python
logging
argparse
black
ExifRead
#opencv-contrib-python
#dautil
#matplotlib